+++ src/components/home/home.tsx
@@ -2,7 +2,7 @@
 import React from 'react'
 import { OurTopics } from './our-topics'
 import { HeroContent } from './hero-content'
-import { LearnInput2 } from '../learn-input/2.0/learn-input2'
+import { LearnInput } from '../learn-input/learn-input'
 
 export const HomePage = () => {
 
@@ -10,7 +10,7 @@ export const HomePage = () => {
         <div className="min-h-screen flex items-center justify-center w-full">
             <div className="flex flex-col items-center gap-8 sm:gap-24 w-full max-w-7xl mx-auto py-8 sm:py-24">
                 <HeroContent />
-                <LearnInput2 />
+                <LearnInput />
                 <OurTopics />
             </div>
         </div>
