+++ src/components/learn-input/loading-bubble.tsx
