+++ src/components/learn-input/options-display.tsx
