+++ src/components/learn-input/show-more-option.tsx
